refactor(state): rename mempool_ports module to mempool_adapters
The module defines adapters (MempoolSourceAdapter, NfsEpochAdapter), not ports —
its own header calls them "the two adapters below". The sibling source_ports.rs
holds an actual consumer bound, so the `_ports` suffix meant two different things.
Rename to mempool_adapters; the ports (traits) live in zaino-mempool/zaino-source.
